communication query
