+++ server/src/main/java/com/omigost/server/model/Communication.java
@@ -4,12 +4,14 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.omigost.server.notification.NotificationService;
 import com.omigost.server.notification.slack.SlackService;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 
 @Data
 @Entity
+@EqualsAndHashCode
 public class Communication {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
