+++ server/src/main/resources/application.properties
@@ -22,3 +22,5 @@ spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults=false
 spring.jpa.hibernate.ddl-auto=update
 spring.output.ansi.enabled=ALWAYS
 logging.level.com.omigost.server=INFO
+
+spring.jpa.show-sql=true
